+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/targets/KotlinMultiplatformJsExtension.kt
@@ -3,7 +3,10 @@ package com.javiersc.hubdle.extensions.kotlin.multiplatform.targets
 import com.javiersc.hubdle.extensions.HubdleDslMarker
 import com.javiersc.hubdle.extensions._internal.state.hubdleState
 import com.javiersc.hubdle.extensions.options.EnableableOptions
+import org.gradle.api.Action
 import org.gradle.api.Project
+import org.jetbrains.kotlin.gradle.targets.js.dsl.KotlinJsBrowserDsl
+import org.jetbrains.kotlin.gradle.targets.js.dsl.KotlinJsNodeDsl
 
 @HubdleDslMarker
 public open class KotlinMultiplatformJsExtension :
@@ -14,13 +17,15 @@ public open class KotlinMultiplatformJsExtension :
     public override val name: String = "js"
 
     @HubdleDslMarker
-    public fun Project.browser() {
-        hubdleState.kotlin.multiplatform.js.browser = true
+    public fun Project.browser(action: Action<KotlinJsBrowserDsl> = Action {}) {
+        hubdleState.kotlin.multiplatform.js.browser.isEnabled = true
+        hubdleState.kotlin.multiplatform.js.browser.action = action
     }
 
     @HubdleDslMarker
-    public fun Project.nodeJs() {
-        hubdleState.kotlin.multiplatform.js.nodeJs = true
+    public fun Project.nodejs(action: Action<KotlinJsNodeDsl> = Action {}) {
+        hubdleState.kotlin.multiplatform.js.nodejs.isEnabled = true
+        hubdleState.kotlin.multiplatform.js.nodejs.action = action
     }
 
     public companion object {